Handle interruptions on nREPL classloader
When nrepl is run in the same process but in a different classloader,
it will receive the same interruption signals that we're handling on
the client side. We need to capture those, so we only handle them
on the client side, and forward the interruption on the nREPL in the
preferred way, via :interrupt.

This has the unfortunate result that we blow away any signal handling
that the nREPL server wants to do, replacing it with a no-op.
Currently the server has just the default Java signal handling, but
this could change. However, this is only a potential problem when the
nREPL server is running in the same process as the client - that's the
only way it can receive the signal.