Read in the proper namespace
So, we will need to have a parallel set of namespaces on the
read side after all. This is for things like `(foo bar), so
that foo and bar will resolve in the proper nses instead of
clojure.core, like they were previously doing. Syntax-quote,
my old nemesis, we meet again.