Use jline for standalone version again
There are huge tradeoffs here. The worst of it, the way I see it, is
requiring us to keep the ugliness of reply.reader.jline and the
JlineInputReader around a bit longer. But that can be mitigated by
eliminating lots of duplication between reply.reader.jline and
simple-jline.

From the user perspective, we also have to choose between reading
multiple forms on the same line vs. decoupling jline and input-reading
via stuff like (read). There may be a way to do both, but I can't think
of one (that also works for multi-line forms).