Standalone REPL: make read-time errors print an exception instead of returning an empty string.
Note: because *e must be set for read-time errors (in addition to
eval-time errors), read-string has been moved into
(eval-state/with-bindings ...)

Fix a bug where empty input lines caused *1 to be set to "".

Fix a bug where tab-completing caused *1 to be set to the tab
completion results.

Add more integration tests for the standalone REPL.