Fix completion-eval to not bind to namespaces that don't exist.
The `reply.eval-modes.nrepl/completion-eval` function would attempt to
bind `*ns*` to the value inside `current-ns`, which starts out as
`reply.eval-modes.nrepl`, a namespace that exists on the client but
usually doesn't exist on the server.

This patch changes it so that it falls back to the current value of
`*ns*` if the namespace in question doesn't exist.