Get rid of backwards-compatibility shims.
There were two .java files included; one that was a backport of a fix
that got added to the PushbackReader in Clojure 1.5, and a main class
which was needed before clojure.main added the -m argument.

The main class was only used from `bin/reply.sh` which relied on a lein1
style lib directory full of jars, so that's deleted as well.

Because technically we're removing a class from the jar file, I've
bumped the version to 0.6.0-SNAPSHOT.

Stop using separate `src/clj` and `src/java` source paths since no
more Java is needed.

Bump parsley version to avoid a warning.